move SubsettableCrossProduct to models.em
